Update FlutterSingleChildScrollViewAbsorbsGesturesExample
An example of how GestureDetection doesn't register in my custom ScrollPhysics ScrollView. The blue ScrollView uses the default ScrollPhysics and GestureDetection works as expected, but the yellow ScrollView on the left uses a custom ScrollPhysics and it seems to only detect gestures when it's in overscroll. Does anyone know what's going on here?!?! Thanks!